LayoutTests/sputnik/Conformance/12_Statement/12.2_Variable_Statement/S12.2_A8_T3-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected token '&&'
S12.2_A8_T3

PASS Expected parsing failure

TEST COMPLETE
